feat: ダッシュボードをGitHub APIリアルタイム連携に変更
- page.tsx を Server Component 化、GitHub API データ取得
- DashboardClient.tsx にUI分離（"use client"）
- github.ts: CI状態・Issues・PRs・リポジトリ活動を自動取得
- ブロッカーセクション → Issues セクション（優先度順、GitHubリンク付き）
- システムタブ: GitHub Actions CI状態をリアルタイム表示
- GITHUB_TOKEN 設定で自動更新（5分キャッシュ）
